requirements.txt
aiogram>=3.4.1
fastapi>=0.110.0
uvicorn>=0.27.1
google-cloud-firestore>=2.15.0
google-cloud-aiplatform>=1.43.0
python-dotenv>=1.0.1
pydantic-settings>=2.2.1
